\name{PGBVS-package}
\alias{PGBVS-package}
\alias{PGBVS}
\docType{package}
\title{
\packageTitle{PGBVS}
}
\description{
\packageDescription{PGBVS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PGBVS}
\packageIndices{PGBVS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PGBVS}

Maintainer: \packageMaintainer{PGBVS}
}
\references{~~ Literature or other references for background information ~~}
\keyword{ package }
\examples{~~ simple examples of the most important functions ~~}
